refactor: move some alias logic out and delay importing
